Only show branding if explictly requested
Previously the branding output (app name, version) was intended
to be shown by the application at all times, but a logic bug
only emitted the branding output when an application error
condition occurred. This commit adds a `branding` flag to
enable toggling the emission of that information when desired.

Because this branding output may not mix well with branding
output from other tools (e.g, atc0005/send2teams), it is not
shown by default.

Other changes:

- README
  - Update README example to reflect use of `branding` flag
    (error output example)
  - Add new flag

- Add `version` field to log output to aid in future
  troubleshooting efforts

refs #13