Update status output to match original plugin
Match the same format used in the original Python 2 plugin.

For reasons I've yet to spend sufficient time to figure
out, the double-quoting used for elements of the "folders"
list is lost when sent by Teams or email notifications. It
is easier to go ahead and just revert the format for now so
it is consistent in each format (console, Teams or email).

refs #16