Explicitly disable use of cgo
As far as I can tell, the v0.2.5 release provided binaries
that are *not* statically linked against glibc and thus
would not be held to LGPL license requirements. AFAICT.

Update Makefile to disable use of cgo for binary generation
in an effort to avoid potential inclusion in the future.

This is likely *not* the last time I'll need to update
the Makefile build options for static binary generation,
but hopefully this will do for now.

refs atc0005/check-mail#94
refs atc0005/mysql2sqlite#48