Add support for escaping Textile formatting chars
Add support to automatically escape or replace known
problematic Textile formatting characters to prevent broken
rendering of email reports in Redmine tickets.

At present only the pipe character is handled, but this work
is intended to allow easily extendeding the map in the future
as other problematic patterns are found.

refs GH-153