Override hardcoded tcp (auto) network selection
Override the `go-imap/client.Dial*` function behavior so that
we have the option to use user-specified network type instead
of relying entirely on correct auto-selection based on the
hardcoded `tcp` named network.

If overridden, the original hardcoded network name and
user-specified network type is emitted at debug log level
for troubleshooting purposes.

This set of changes is considered an extension of the original
hotfix applied by filtering DNS query results to one of
IPv4-only, IPv6-only or "all".

refs GH-172