Add lsimap troubleshooting tool
The lsimap binary is intended to provide a simple CLI tool to
list advertised IMAP server capabilities.

As part of implementing this tool some minor tweaks were made
to the config and mbxs packages to mute some existing log
messages by default. Equivalent log output was added to the
check_imap_mailbox and list-emails tools to compensate for
the change.

README updated with examples for lsimap tool.